renamed and broke out calculator from temperature
